test: stub services package before imports
